Move workout modal to a separate component that way it is reusable. Add workout creation to calendar. remove workout creation from workout page for now
