+++ src/FitnessTracker.Web/src/pages/Workouts.jsx
@@ -1,5 +1,6 @@
 import { useState, useEffect } from 'react';
 import { workoutsApi, exerciseDefinitionsApi } from '../services/api';
+import WorkoutModal from '../components/WorkoutModal';
 
 function Workouts() {
     const [workouts, setWorkouts] = useState([]);
@@ -7,16 +8,6 @@ function Workouts() {
     const [loading, setLoading] = useState(true);
     const [showModal, setShowModal] = useState(false);
     const [editingWorkout, setEditingWorkout] = useState(null);
-    const [formData, setFormData] = useState({
-        title: '',
-        description: '',
-        durationMinutes: 30,
-        workoutDate: new Date().toISOString().split('T')[0],
-        exercises: [],
-    });
-    const [collapsedExercises, setCollapsedExercises] = useState(new Set());
-    const usedExerciseDefinitionIds = formData.exercises.map(e => e.exerciseDefinitionId);
-    const unusedExerciseDefinitions = definitions.filter(d => !usedExerciseDefinitionIds.includes(d.id));
 
     useEffect(() => {
         loadData();
@@ -25,7 +16,6 @@ function Workouts() {
     const loadData = async () => {
         try {
             setLoading(true);
-            //Get the list of workouts and exercise definitions on component load
             const [workoutsData, definitionsData] = await Promise.all([
                 workoutsApi.getAll(),
                 exerciseDefinitionsApi.getAll()
@@ -39,7 +29,6 @@ function Workouts() {
         }
     };
 
-    //Use this function to hot-reload workouts on actions like creating/deleting
     const loadWorkouts = async () => {
         try {
             const data = await workoutsApi.getAll();
@@ -50,229 +39,13 @@ function Workouts() {
     };
 
     const handleOpenModal = (workout = null) => {
-        if (workout) {
-            setEditingWorkout(workout);
-            const exercises = workout.exercises.map(ex => ({
-                ...ex,
-                sets: ex.sets.map(s => ({ ...s }))
-            })) || [];
-            setFormData({
-                title: workout.title,
-                description: workout.description || '',
-                durationMinutes: workout.durationMinutes,
-                workoutDate: workout.workoutDate.split('T')[0],
-                exercises,
-            });
-            // Initialize all exercises as collapsed
-            setCollapsedExercises(new Set(exercises.map((_, idx) => idx)));
-        } else {
-            setEditingWorkout(null);
-            setFormData({
-                title: '',
-                description: '',
-                durationMinutes: 30,
-                workoutDate: new Date().toISOString().split('T')[0],
-                exercises: [],
-            });
-            setCollapsedExercises(new Set());
-        }
+        setEditingWorkout(workout);
         setShowModal(true);
     };
 
     const handleCloseModal = () => {
         setShowModal(false);
         setEditingWorkout(null);
-    };
-
-    const handleInputChange = (e) => {
-        const { name, value } = e.target;
-        setFormData((prev) => ({
-            ...prev,
-            [name]: name === 'durationMinutes' ? parseInt(value) || 0 : value,
-        }));
-    };
-
-    const handleAddExercise = () => {
-        const usedIds = formData.exercises.map(e => e.exerciseDefinitionId);
-        const unusedDefs = definitions.filter(def => !usedIds.includes(def.id));
-        //Dont add if no exercises left to add, prevent button from doing anything. Need to update the UI to disable the button
-        if (unusedDefs.length === 0) {
-            alert('No more exercises to add!');
-            return;
-        }
-        setFormData((prev) => ({
-            ...prev,
-            exercises: [
-                ...prev.exercises,
-                {
-                    exerciseDefinitionId: unusedDefs[0].id,
-                    notes: '',
-                    sets: [{ setNumber: 1, reps: 10, weight: null }]
-                },
-            ],
-        }));
-    };
-
-    const handleExerciseChange = (index, field, value) => {
-        setFormData((prev) => ({
-            ...prev,
-            exercises: prev.exercises.map((ex, i) =>
-                i === index
-                    ? {
-                        ...ex,
-                        [field]: value,
-                    }
-                    : ex
-            ),
-        }));
-    };
-
-    const handleAddSet = (exerciseIndex) => {
-        setFormData((prev) => {
-            const newExercises = [...prev.exercises];
-            const exercise = { ...newExercises[exerciseIndex] };
-            const sets = [...exercise.sets];
-            const lastSet = sets[sets.length - 1];
-
-            sets.push({
-                setNumber: sets.length + 1,
-                reps: lastSet?.reps || 10,
-                weight: lastSet?.weight || null,
-            });
-
-            exercise.sets = sets;
-            newExercises[exerciseIndex] = exercise;
-            return { ...prev, exercises: newExercises };
-        });
-    };
-
-    const handleSetChange = (exerciseIndex, setIndex, field, value) => {
-        setFormData((prev) => {
-            const newExercises = [...prev.exercises];
-            const exercise = { ...newExercises[exerciseIndex] };
-            const sets = [...exercise.sets];
-
-            sets[setIndex] = {
-                ...sets[setIndex],
-                [field]: value === '' ? null : (field === 'weight' ? parseFloat(value) : parseInt(value)),
-            };
-
-            exercise.sets = sets;
-            newExercises[exerciseIndex] = exercise;
-            return { ...prev, exercises: newExercises };
-        });
-    };
-
-    const handleRemoveSet = (exerciseIndex, setIndex) => {
-        setFormData((prev) => {
-            const newExercises = [...prev.exercises];
-            const exercise = { ...newExercises[exerciseIndex] };
-            exercise.sets = exercise.sets.filter((_, i) => i !== setIndex).map((s, i) => ({ ...s, setNumber: i + 1 }));
-            newExercises[exerciseIndex] = exercise;
-            return { ...prev, exercises: newExercises };
-        });
-    };
-
-    const handleRemoveExercise = (index) => {
-        setFormData((prev) => ({
-            ...prev,
-            exercises: prev.exercises.filter((_, i) => i !== index),
-        }));
-    };
-
-    const toggleExerciseCollapse = (index) => {
-        setCollapsedExercises(prev => {
-            const newSet = new Set(prev);
-            if (newSet.has(index)) {
-                newSet.delete(index);
-            } else {
-                newSet.add(index);
-            }
-            return newSet;
-        });
-    };
-
-    const moveExerciseUp = (index) => {
-        if (index === 0) return;
-        setFormData((prev) => {
-            const newExercises = [...prev.exercises];
-            [newExercises[index - 1], newExercises[index]] = [newExercises[index], newExercises[index - 1]];
-            return { ...prev, exercises: newExercises };
-        });
-
-        // Update collapsed state
-        setCollapsedExercises(prev => {
-            const newSet = new Set();
-            prev.forEach(i => {
-                if (i === index) newSet.add(index - 1);
-                else if (i === index - 1) newSet.add(index);
-                else newSet.add(i);
-            });
-            return newSet;
-        });
-    };
-
-    const moveExerciseDown = (index) => {
-        if (index >= formData.exercises.length - 1) return;
-        setFormData((prev) => {
-            const newExercises = [...prev.exercises];
-            [newExercises[index], newExercises[index + 1]] = [newExercises[index + 1], newExercises[index]];
-            return { ...prev, exercises: newExercises };
-        });
-
-        // Update collapsed state
-        setCollapsedExercises(prev => {
-            const newSet = new Set();
-            prev.forEach(i => {
-                if (i === index) newSet.add(index + 1);
-                else if (i === index + 1) newSet.add(index);
-                else newSet.add(i);
-            });
-            return newSet;
-        });
-    };
-
-    const toggleCollapseAll = () => {
-        // If all are collapsed, expand all. Otherwise, collapse all.
-        const allCollapsed = formData.exercises.every((_, idx) => collapsedExercises.has(idx));
-
-        if (allCollapsed) {
-            // Expand all
-            setCollapsedExercises(new Set());
-        } else {
-            // Collapse all
-            setCollapsedExercises(new Set(formData.exercises.map((_, idx) => idx)));
-        }
-    };
-
-    const handleSubmit = async (e) => {
-        e.preventDefault();
-        try {
-            // Clean up data for submission
-            const submissionData = {
-                ...formData,
-                exercises: formData.exercises.map(ex => ({
-                    exerciseDefinitionId: parseInt(ex.exerciseDefinitionId),
-                    notes: ex.notes,
-                    sets: ex.sets.map(s => ({
-                        setNumber: s.setNumber,
-                        reps: s.reps != null ? parseInt(s.reps) : null,
-                        weight: s.weight != null ? Math.round(parseFloat(s.weight) * 100) / 100 : null
-                    }))
-                }))
-            };
-
-            if (editingWorkout) {
-                await workoutsApi.update(editingWorkout.id, submissionData);
-            } else {
-                await workoutsApi.create(submissionData);
-            }
-            handleCloseModal();
-            loadWorkouts();
-        } catch (error) {
-            console.error('Failed to save workout:', error);
-            alert(error.message || 'Failed to save workout. Please check your inputs.');
-        }
     };
 
     const handleDelete = async (id) => {
@@ -298,9 +71,7 @@ function Workouts() {
         <div className="fade-in">
             <div className="page-header">
                 <h1 className="page-title">Workouts</h1>
-                <button className="btn btn-primary" onClick={() => handleOpenModal()}>
-                    ➕ Log Workout
-                </button>
+                {/* Log Workout button removed - use Calendar to create workouts */}
             </div>
 
             {workouts.length === 0 ? (
@@ -308,14 +79,7 @@ function Workouts() {
                     <div className="empty-state">
                         <div className="empty-state-icon">🏋️</div>
                         <h3>No workouts yet</h3>
-                        <p>Start logging your workouts to track your progress!</p>
-                        <button
-                            className="btn btn-primary"
-                            onClick={() => handleOpenModal()}
-                            style={{ marginTop: '1rem' }}
-                        >
-                            Log Your First Workout
-                        </button>
+                        <p>Go to the Calendar page to log your first workout!</p>
                     </div>
                 </div>
             ) : (
@@ -335,14 +99,12 @@ function Workouts() {
                                     </p>
                                 </div>
                                 <div style={{ display: 'flex', gap: '0.5rem' }}>
-                                    {/* Edit button next to workout */}
                                     <button
                                         className="btn btn-secondary btn-sm"
                                         onClick={() => handleOpenModal(workout)}
                                     >
                                         ✏️
                                     </button>
-                                    {/* Delete button next to workout */}
                                     <button
                                         className="btn btn-danger btn-sm"
                                         onClick={() => handleDelete(workout.id)}
@@ -369,7 +131,6 @@ function Workouts() {
                                 </div>
                             </div>
 
-                            {/* Personal Records */}
                             {workout.exercises.length > 0 && (
                                 <div className="exercise-list">
                                     {workout.exercises.map((exercise) => (
@@ -404,297 +165,15 @@ function Workouts() {
             )}
 
             {/* Workout Modal */}
-            {showModal && (
-                <div className="modal-overlay" onClick={handleCloseModal}>
-                    <div className="modal" onClick={(e) => e.stopPropagation()}>
-                        <div className="modal-header">
-                            <h2 className="modal-title">
-                                {editingWorkout ? 'Edit Workout' : 'Log Workout'}
-                            </h2>
-                            <button
-                                className="btn btn-secondary btn-icon"
-                                onClick={handleCloseModal}
-                            >
-                                ✕
-                            </button>
-                        </div>
-
-                        <form onSubmit={handleSubmit}>
-                            <div className="modal-body">
-                                <div className="form-group">
-                                    <label className="form-label">Workout Title</label>
-                                    <input
-                                        type="text"
-                                        name="title"
-                                        className="form-input"
-                                        value={formData.title}
-                                        onChange={handleInputChange}
-                                        placeholder="e.g., Upper Body Strength"
-                                        required
-                                    />
-                                </div>
-
-                                <div className="grid grid-2">
-                                    <div className="form-group">
-                                        <label className="form-label">Date</label>
-                                        <input
-                                            type="date"
-                                            name="workoutDate"
-                                            className="form-input"
-                                            value={formData.workoutDate}
-                                            onChange={handleInputChange}
-                                            required
-                                        />
-                                    </div>
-                                    <div className="form-group">
-                                        <label className="form-label">Duration (minutes)</label>
-                                        <input
-                                            type="number"
-                                            name="durationMinutes"
-                                            className="form-input"
-                                            value={formData.durationMinutes}
-                                            onChange={handleInputChange}
-                                            min="1"
-                                            required
-                                        />
-                                    </div>
-                                </div>
-
-                                <div className="form-group">
-                                    <label className="form-label">Notes (optional)</label>
-                                    <textarea
-                                        name="description"
-                                        className="form-textarea"
-                                        value={formData.description}
-                                        onChange={handleInputChange}
-                                        placeholder="How did the workout go?"
-                                    />
-                                </div>
-
-                                <div className="form-group">
-                                    <div style={{ display: 'flex', justifyContent: 'space-between', alignItems: 'center', marginBottom: '1rem' }}>
-                                        <label className="form-label" style={{ margin: 0 }}>
-                                            Exercises
-                                        </label>
-                                        {/* Collapse all button */}
-                                        <div style={{ display: 'flex', gap: '0.5rem' }}>
-                                            {formData.exercises.length > 0 && (
-                                                <button
-                                                    type="button"
-                                                    className="btn btn-secondary btn-sm"
-                                                    onClick={toggleCollapseAll}
-                                                    title={formData.exercises.every((_, idx) => collapsedExercises.has(idx)) ? "Expand all sets" : "Collapse all sets"}
-                                                >
-                                                    {formData.exercises.every((_, idx) => collapsedExercises.has(idx)) ? '📂 Expand All' : '📁 Collapse All'}
-                                                </button>
-                                            )}
-                                            <button
-                                                type="button"
-                                                className="btn btn-secondary btn-sm"
-                                                onClick={handleAddExercise}
-                                                disabled={unusedExerciseDefinitions.length === 0}
-                                            >
-                                                ➕ Add Exercise
-                                            </button>
-                                        </div>
-                                    </div>
-
-                                    <div className="exercise-list">
-                                        {formData.exercises.map((exercise, index) => {
-                                            const isCollapsed = collapsedExercises.has(index);
-                                            const selectedDef = definitions.find(d => d.id == exercise.exerciseDefinitionId);
-                                            // Show current selection + unused exercises for this dropdown
-                                            const availableForThisRow = definitions.filter(d =>
-                                                d.id == exercise.exerciseDefinitionId ||
-                                                !formData.exercises.some((ex, i) => i !== index && ex.exerciseDefinitionId == d.id)
-                                            );
-
-                                            return (
-                                                <div
-                                                    key={index}
-                                                    style={{
-                                                        background: 'var(--bg-secondary)',
-                                                        padding: '1rem',
-                                                        borderRadius: 'var(--radius-md)',
-                                                        marginBottom: '0.5rem',
-                                                        border: '1px solid var(--border-color)',
-                                                    }}
-                                                >
-                                                    {/* Exercise Header with Reorder Controls */}
-                                                    <div style={{ display: 'flex', gap: '0.5rem', marginBottom: '0.75rem', alignItems: 'center' }}>
-                                                        {/* Reorder buttons */}
-                                                        <div style={{ display: 'flex', flexDirection: 'column', gap: '2px' }}>
-                                                            <button
-                                                                type="button"
-                                                                className="btn btn-secondary"
-                                                                onClick={() => moveExerciseUp(index)}
-                                                                disabled={index === 0}
-                                                                style={{
-                                                                    padding: '2px 8px',
-                                                                    fontSize: '0.7rem',
-                                                                    minWidth: '28px',
-                                                                    opacity: index === 0 ? 0.3 : 1,
-                                                                    cursor: index === 0 ? 'not-allowed' : 'pointer'
-                                                                }}
-                                                                title="Move up"
-                                                            >
-                                                                ▲
-                                                            </button>
-                                                            <button
-                                                                type="button"
-                                                                className="btn btn-secondary"
-                                                                onClick={() => moveExerciseDown(index)}
-                                                                disabled={index >= formData.exercises.length - 1}
-                                                                style={{
-                                                                    padding: '2px 8px',
-                                                                    fontSize: '0.7rem',
-                                                                    minWidth: '28px',
-                                                                    opacity: index >= formData.exercises.length - 1 ? 0.3 : 1,
-                                                                    cursor: index >= formData.exercises.length - 1 ? 'not-allowed' : 'pointer'
-                                                                }}
-                                                                title="Move down"
-                                                            >
-                                                                ▼
-                                                            </button>
-                                                        </div>
-
-                                                        {/* Exercise dropdown */}
-                                                        <select
-                                                            className="form-input"
-                                                            value={exercise.exerciseDefinitionId}
-                                                            onChange={(e) =>
-                                                                handleExerciseChange(index, 'exerciseDefinitionId', e.target.value)
-                                                            }
-                                                            required
-                                                            style={{ flex: 1 }}
-                                                        >
-                                                            {availableForThisRow.map(def => (
-                                                                <option key={def.id} value={def.id}>
-                                                                    {def.name} ({def.primaryMuscleGroup})
-                                                                </option>
-                                                            ))}
-                                                        </select>
-
-                                                        {/* Delete button */}
-                                                        <button
-                                                            type="button"
-                                                            className="btn btn-danger btn-icon"
-                                                            onClick={() => handleRemoveExercise(index)}
-                                                            title="Remove exercise"
-                                                        >
-                                                            ✕
-                                                        </button>
-                                                    </div>
-
-                                                    {/* Sets Section - Collapsible */}
-                                                    <div className="sets-container" style={{ marginTop: '0.5rem' }}>
-                                                        <div
-                                                            style={{
-                                                                display: 'flex',
-                                                                justifyContent: 'space-between',
-                                                                alignItems: 'center',
-                                                                marginBottom: '0.5rem',
-                                                                cursor: 'pointer',
-                                                                userSelect: 'none'
-                                                            }}
-                                                            onClick={() => toggleExerciseCollapse(index)}
-                                                        >
-                                                            <div style={{ display: 'flex', alignItems: 'center', gap: '0.5rem' }}>
-                                                                <span style={{ fontSize: '0.9rem', fontWeight: 'bold' }}>
-                                                                    {isCollapsed ? '▶' : '▼'} Sets
-                                                                </span>
-                                                                <span style={{ fontSize: '0.75rem', color: 'var(--text-muted)' }}>
-                                                                    ({exercise.sets.length})
-                                                                </span>
-                                                            </div>
-                                                            {!isCollapsed && (
-                                                                <button
-                                                                    type="button"
-                                                                    className="btn btn-secondary btn-xs"
-                                                                    onClick={(e) => {
-                                                                        e.stopPropagation();
-                                                                        handleAddSet(index);
-                                                                    }}
-                                                                >
-                                                                    + Add Set
-                                                                </button>
-                                                            )}
-                                                        </div>
-
-                                                        {!isCollapsed && (
-                                                            <div style={{
-                                                                animation: 'fadeIn 0.2s ease',
-                                                                overflow: 'hidden'
-                                                            }}>
-                                                                {exercise.sets.map((set, setIndex) => (
-                                                                    <div key={setIndex} className="grid grid-3" style={{ gap: '0.5rem', marginBottom: '0.5rem', alignItems: 'center' }}>
-                                                                        <div className="set-label">#{set.setNumber}</div>
-                                                                        <input
-                                                                            type="number"
-                                                                            className="form-input btn-sm"
-                                                                            placeholder="Reps"
-                                                                            value={set.reps || ''}
-                                                                            onChange={(e) => handleSetChange(index, setIndex, 'reps', e.target.value)}
-                                                                            required
-                                                                        />
-                                                                        <div style={{ display: 'flex', gap: '0.25rem', alignItems: 'center' }}>
-                                                                            <input
-                                                                                type="number"
-                                                                                step="0.5"
-                                                                                className="form-input btn-sm"
-                                                                                placeholder="lbs"
-                                                                                value={set.weight === null ? '' : set.weight}
-                                                                                onChange={(e) => handleSetChange(index, setIndex, 'weight', e.target.value)}
-                                                                                max="2000"
-                                                                            />
-                                                                            <button
-                                                                                type="button"
-                                                                                className="btn btn-danger btn-xs"
-                                                                                onClick={() => handleRemoveSet(index, setIndex)}
-                                                                                disabled={exercise.sets.length === 1}
-                                                                            >
-                                                                                ✕
-                                                                            </button>
-                                                                        </div>
-                                                                    </div>
-                                                                ))}
-                                                            </div>
-                                                        )}
-                                                    </div>
-
-                                                    {/* Exercise Notes */}
-                                                    <div style={{ marginTop: '0.75rem' }}>
-                                                        <input
-                                                            type="text"
-                                                            className="form-input btn-sm"
-                                                            placeholder="Exercise notes..."
-                                                            value={exercise.notes || ''}
-                                                            onChange={(e) => handleExerciseChange(index, 'notes', e.target.value)}
-                                                        />
-                                                    </div>
-                                                </div>
-                                            );
-                                        })}
-                                    </div>
-                                </div>
-                            </div>
-
-                            <div className="modal-footer">
-                                <button
-                                    type="button"
-                                    className="btn btn-secondary"
-                                    onClick={handleCloseModal}
-                                >
-                                    Cancel
-                                </button>
-                                <button type="submit" className="btn btn-primary">
-                                    {editingWorkout ? 'Save Changes' : 'Log Workout'}
-                                </button>
-                            </div>
-                        </form>
-                    </div>
-                </div>
-            )}
+            <WorkoutModal
+                isOpen={showModal}
+                onClose={handleCloseModal}
+                onSave={loadWorkouts}
+                workout={editingWorkout}
+                initialDate={new Date().toISOString().split('T')[0]}
+                definitions={definitions}
+                templateData={null}
+            />
         </div>
     );
 }
